pyproject.toml
[project]
name = "babbel-tux"
version = "0.1.0"
description = "Telegram bot for the LUG-Taunus group written in Python"
authors = [{ name = "Philip Schömig", email = "philip.schoemig@posteo.de" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
babbel-tux = "babbel_tux.main:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2,<9",
    "pytest-cov>=7.0.0,<8",
    "ruff>=0.14.3",
    "tox>=4.30.2,<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D401"]
